+++ Plugin.BluetoothLE.Abstractions/AbstractGattService.cs
@@ -18,5 +18,6 @@ namespace Plugin.BluetoothLE
         public bool IsPrimary { get; }
         public virtual string Description => Dictionaries.GetServiceDescription(this.Uuid);
         public abstract IObservable<IGattCharacteristic> WhenCharacteristicDiscovered();
+        public abstract IObservable<IGattCharacteristic> GetKnownCharacteristics(params Guid[] characteristicIds);
     }
 }